# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  oauthDTO.Response:
    properties:
      is_active:
        type: boolean
      message:
        type: string
      oauth_token:
        type: string
      oauth_url:
        type: string
      status:
        type: string
    type: object
  userDTO.GenericResponse:
    properties:
      data: {}
      error:
        type: string
      status:
        type: string
    type: object
  userDTO.LoginResponse:
    properties:
      status:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/userDTO.User'
    type: object
...
  userDTO.UpdatePassword:
    properties:
      new_password1:
        type: string
      new_password2:
        type: string
      old_password:
        type: string
    required:
    - new_password1
    - new_password2
    - old_password
    type: object
  userDTO.User:
    properties:
      account:
        $ref: '#/definitions/usermodel.AccountType'
      country:
        $ref: '#/definitions/usermodel.RegionType'
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_disabled:
        type: boolean
      is_verified:
        type: boolean
      last_name:
        type: string
      oauth:
        $ref: '#/definitions/oauthDTO.Response'
      updated_at:
        type: string
      username:
        type: string
    type: object
  userDTO.UserCreate:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      region:
        $ref: '#/definitions/usermodel.RegionType'
      username:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - region
    - username
    type: object
  userDTO.UserLogin:
    properties:
      otp:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  usermodel.AccountType:
    enum:
    - User
    - Seller
    - Admin
    type: string
    x-enum-varnames:
    - AccountTypeNormal
    - AccountTypeSeller
    - AccountTypeAdmin
  usermodel.RegionType:
    enum:
    - Northern America
    - Central America
    - Caribbean
    - Northern South America
    - Southern South America
    - Western South America
    - Eastern South America
    - Scandinavia
    - Southern Europe
    - Western Europe
    - Eastern Europe
    - Central Europe
    - Middle East
    - Central Asia
    - Eastern Asia
    - Southern Asia
    - Southeastern Asia
    - Siberia
    - Northern Africa
    - Western Africa
    - Central Africa
    - Horn of Africa
    - Southern Africa
    - Australia and New Zealand
    - Melanesia
    - Micronesia
    - Polynesia
    type: string
    x-enum-varnames:
    - RegionTypeNorthernAmerica
    - RegionTypeCentralAmerica
    - RegionTypeCaribbean
    - NorthernSouthAmerica
    - SouthernSouthAmerica
    - WesternSouthAmerica
    - EasternSouthAmerica
    - RegionTypeScandinavia
    - RegionTypeSouthernEurope
    - RegionTypeWesternEurope
    - RegionTypeEasternEurope
    - RegionTypeCentralEurope
    - RegionTypeMiddleEast
    - RegionTypeCentralAsia
    - RegionTypeEasternAsia
    - RegionTypeSouthernAsia
    - RegionTypeSoutheasternAsia
    - RegionTypeSiberia
    - RegionTypeNorthernAfrica
    - RegionTypeWesternAfrica
    - RegionTypeCentralAfrica
    - RegionTypeHornOfAfrica
    - RegionTypeSouthernAfrica
    - AustraliaAndNewZealand
    - Melanesia
    - Micronesia
    - Polynesia
info:
  contact: {}
  description: API for Rasta
  title: Rasta API
  version: "1.0"
paths:
  /admin/users:
    get:
      consumes:
      - application/json
      description: Get a list of users with pagination support
      parameters:
      - default: 10
        description: Number of users per page
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
      summary: Get users with pagination
      tags:
      - Users
  /admin/users/count:
    get:
      consumes:
      - application/json
      description: Retrieve the total count of users in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
      summary: Get the total number of users
      tags:
      - Users
  /admin/users/id/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
      summary: Delete a user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retrieve user details by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
      summary: Get user by ID
      tags:
      - Users
...
  /users/{username}:
    get:
      consumes:
      - application/json
      description: Retrieve user details by their username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
      summary: Get user by username
      tags:
      - Users
  /users/{username}/update-password:
    put:
      consumes:
      - application/json
      description: Updates the password for the currently authenticated user
      parameters:
      - description: Password update payload
        in: body
        name: updatePassword
        required: true
        schema:
          $ref: '#/definitions/userDTO.UpdatePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
      summary: Update user password
      tags:
      - Users
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: User login payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/userDTO.UserLogin'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/userDTO.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
      summary: User login
      tags:
      - Users
...
  /users/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account and send a verification OTP email
      parameters:
      - description: User creation payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/userDTO.UserCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
      summary: Create a new user
      tags:
      - Users
swagger: "2.0"